finish project 2

+++ Project2-Quantifying_Alzheimer_Disease_Progression/section2/src/data_prep/HippocampusDatasetLoader.py
@@ -39,7 +39,7 @@ def LoadHippocampusData(root_dir, y_shape, z_shape):
         label, _ = load(os.path.join(label_dir, f))
 
         # TASK: normalize all images (but not labels) so that values are in [0..1] range
-        image = image / np.max(image)
+        image = image.astype(np.single) / np.max(image)
 
         # We need to reshape data since CNN tensors that represent minibatches
         # in our case will be stacks of slices and stacks need to be of the same size.
